pyproject.toml
[tool.poetry]
name = "nitpick"
version = "0.1.0"
description = ""
authors = ["cashmere53 <15913755+cashmere53@users.noreply.github.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py", rev = "master"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.971"
pyproject-flake8 = "^0.0.1a5"
flake8-bugbear = "^22.7.1"


[tool.poetry.scripts]
nitpick = "nitpick.__main__:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ["py310"]
line-length = 120


[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
known_local_folder = ["config",]


[tool.flake8]
max-line-length = 120
max-complexity = 10
ignore = "E203,E266,W503,"
exclude = [
    ".venv/",
    ".mypy_cache/",
    ".vscode/"
]


[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true


[tool.pytest.ini_options]
testpaths = ["tests",]
filterwarnings = ["ignore::DeprecationWarning",]
